+++ predicators/ground_truth_models/coffee/nsrts.py
@@ -69,9 +69,10 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
             LiftedAtom(HandEmpty, [robot]),
         }
         ignore_effects: Set[Predicate] = set()
-        move_to_twist_jug_nsrt = NSRT("MoveToTwistJug", parameters, preconditions,
-                                    add_effects, delete_effects, ignore_effects,
-                                    option, option_vars, null_sampler)
+        move_to_twist_jug_nsrt = NSRT("MoveToTwistJug", parameters,
+                                      preconditions, add_effects,
+                                      delete_effects, ignore_effects, option,
+                                      option_vars, null_sampler)
         nsrts.add(move_to_twist_jug_nsrt)
 
         # TwistJug
@@ -93,14 +94,14 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         ignore_effects = set()
 
         def twist_jug_sampler(state: State, goal: Set[GroundAtom],
-                            rng: np.random.Generator,
-                            objs: Sequence[Object]) -> Array:
+                              rng: np.random.Generator,
+                              objs: Sequence[Object]) -> Array:
             del state, goal, objs  # unused
             return np.array(rng.uniform(-1, 1, size=(1, )), dtype=np.float32)
 
-        twist_jug_nsrt = NSRT("TwistJug", parameters, preconditions, add_effects,
-                            delete_effects, ignore_effects, option, option_vars,
-                            twist_jug_sampler)
+        twist_jug_nsrt = NSRT("TwistJug", parameters, preconditions,
+                              add_effects, delete_effects, ignore_effects,
+                              option, option_vars, twist_jug_sampler)
         nsrts.add(twist_jug_nsrt)
 
         # PickJugFromTable
@@ -122,9 +123,9 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         }
         ignore_effects = set()
         pick_jug_from_table_nsrt = NSRT("PickJugFromTable", parameters,
-                                        preconditions, add_effects, delete_effects,
-                                        ignore_effects, option, option_vars,
-                                        null_sampler)
+                                        preconditions, add_effects,
+                                        delete_effects, ignore_effects, option,
+                                        option_vars, null_sampler)
         nsrts.add(pick_jug_from_table_nsrt)
 
         # PlaceJugInMachine
@@ -146,9 +147,9 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         }
         ignore_effects = set()
         place_jug_in_machine_nsrt = NSRT("PlaceJugInMachine", parameters,
-                                        preconditions, add_effects,
-                                        delete_effects, ignore_effects, option,
-                                        option_vars, null_sampler)
+                                         preconditions, add_effects,
+                                         delete_effects, ignore_effects,
+                                         option, option_vars, null_sampler)
         nsrts.add(place_jug_in_machine_nsrt)
 
         # TurnMachineOn
@@ -170,8 +171,9 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         delete_effects = set()
         ignore_effects = set()
         turn_machine_on_nsrt = NSRT("TurnMachineOn", parameters, preconditions,
-                                    add_effects, delete_effects, ignore_effects,
-                                    option, option_vars, null_sampler)
+                                    add_effects, delete_effects,
+                                    ignore_effects, option, option_vars,
+                                    null_sampler)
         nsrts.add(turn_machine_on_nsrt)
 
         # PickJugFromMachine
@@ -196,9 +198,9 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         }
         ignore_effects = set()
         pick_jug_from_machine_nsrt = NSRT("PickJugFromMachine", parameters,
-                                        preconditions, add_effects,
-                                        delete_effects, ignore_effects, option,
-                                        option_vars, null_sampler)
+                                          preconditions, add_effects,
+                                          delete_effects, ignore_effects,
+                                          option, option_vars, null_sampler)
         nsrts.add(pick_jug_from_machine_nsrt)
 
         # PourFromNowhere
@@ -222,9 +224,10 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
             LiftedAtom(NotAboveCup, [robot, jug]),
         }
         ignore_effects = set()
-        pour_from_nowhere_nsrt = NSRT("PourFromNowhere", parameters, preconditions,
-                                    add_effects, delete_effects, ignore_effects,
-                                    option, option_vars, null_sampler)
+        pour_from_nowhere_nsrt = NSRT("PourFromNowhere", parameters,
+                                      preconditions, add_effects,
+                                      delete_effects, ignore_effects, option,
+                                      option_vars, null_sampler)
         nsrts.add(pour_from_nowhere_nsrt)
 
         # PourFromOtherCup
@@ -253,9 +256,9 @@ class CoffeeGroundTruthNSRTFactory(GroundTruthNSRTFactory):
         }
         ignore_effects = set()
         pour_from_other_cup_nsrt = NSRT("PourFromOtherCup", parameters,
-                                        preconditions, add_effects, delete_effects,
-                                        ignore_effects, option, option_vars,
-                                        null_sampler)
+                                        preconditions, add_effects,
+                                        delete_effects, ignore_effects, option,
+                                        option_vars, null_sampler)
         nsrts.add(pour_from_other_cup_nsrt)
 
         return nsrts
