+++ predicators/ground_truth_models/satellites/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for satellites environment and variants."""
+
+from .nsrts import SatellitesGroundTruthNSRTFactory
+
+__all__ = ["SatellitesGroundTruthNSRTFactory"]
